trochu struktůry do děch fajlů
